pyproject.toml
[project]
name = "meme-service"
version = "0.1.0"
description = "Checks for reposts"
readme = "README.md"
dependencies = [
    "aio-pika==9.5.4",
    "clip",
    "faststream==0.5.34",
    "ftfy==6.3.1",
    "marshmallow-dataclass==8.7.1",
    "packaging==24.2",
    "pika==1.3.2",
    "pillow==11.1.0",
    "pydantic==2.10.6",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.3",
    "qdrant-client==1.13.2",
    "regex==2024.11.6",
    "testcontainers==4.9.1",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tqdm==4.67.1",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
